fix(root): clear GIT_DIR in pre-commit as well as pre-push
Git exports GIT_DIR into pre-commit exactly as it does into pre-push, and
you cannot push from a worktree without committing from one first. The two
commands pre-commit runs happen to survive it — gitleaks and lint-staged
both drive the git CLI, which resolves the worktree pointer correctly, and
a commit from a worktree completes with it set, lint-staged's
stash-and-restore included.

That is a property of those two commands rather than of the hook, and it
lapses the moment either changes or a third joins them; turbo is the
standing proof that a git-shelling tool can take the pointer badly. Leaving
it uncleared and asserting in a comment that the tools cope would put the
hook's correctness in third-party internals nothing here can check.

Clearing it costs nothing: with no GIT_DIR set, git discovers the
repository from the working directory, and hooks already run from the top
of the working tree — the same repository in a normal clone, the same
worktree in a linked one.

The test file covers both hooks now, so it is named for them.
